+++ tests/auxiliary/test_wikidata.py
@@ -4,16 +4,16 @@ from pytest import MonkeyPatch
 
 from mex.backend.auxiliary import wikidata
 from mex.common.models import ExtractedPrimarySource
-from mex.common.primary_source.helpers import get_all_extracted_primary_sources
-from mex.common.testing import Joker
 
 
 @pytest.mark.usefixtures("mocked_wikidata")
 def test_search_organization_in_wikidata_mocked(
-    client_with_api_key_read_permission: TestClient, monkeypatch: MonkeyPatch
+    client_with_api_key_read_permission: TestClient,
+    monkeypatch: MonkeyPatch,
+    extracted_primary_sources: dict[str, ExtractedPrimarySource],
 ) -> None:
     def extracted_primary_source_wikidata() -> ExtractedPrimarySource:
-        return get_all_extracted_primary_sources()["wikidata"]
+        return extracted_primary_sources["wikidata"]
 
     monkeypatch.setattr(
         wikidata, "extracted_primary_source_wikidata", extracted_primary_source_wikidata
@@ -27,7 +27,36 @@ def test_search_organization_in_wikidata_mocked(
         "items": [
             {
                 "$type": "ExtractedOrganization",
-                "alternativeName": Joker(),
+                "alternativeName": [
+                    {
+                        "language": None,
+                        "value": "alias_en_1",
+                    },
+                    {
+                        "language": None,
+                        "value": "alias_en_2",
+                    },
+                    {
+                        "language": None,
+                        "value": "alias_en_3",
+                    },
+                    {
+                        "language": None,
+                        "value": "alias_en_4",
+                    },
+                    {
+                        "language": None,
+                        "value": "alias_de_1",
+                    },
+                    {
+                        "language": None,
+                        "value": "alias_de_2",
+                    },
+                    {
+                        "language": None,
+                        "value": "alias_de_3",
+                    },
+                ],
                 "geprisId": [],
                 "gndId": [],
                 "hadPrimarySource": "djbNGb5fLgYHFyMh3fZE2g",
@@ -40,8 +69,8 @@ def test_search_organization_in_wikidata_mocked(
                 ],
                 "rorId": ["https://ror.org/01k5qnb77"],
                 "shortName": [
-                    {"language": "de", "value": "RKI"},
                     {"language": "en", "value": "RKI"},
+                    {"language": "de", "value": "RKI"},
                 ],
                 "stableTargetId": "ga6xh6pgMwgq7DC7r6Wjqg",
                 "viafId": [],