+++ tests/graph/test_connector.py
@@ -192,8 +192,8 @@ def test_on_commit_backoff(monkeypatch: MonkeyPatch) -> None:
         check_connectivity_and_authentication,
     )
 
-    connector = GraphConnector.get()
-    event = BackoffDetails(args=(connector,))
+    graph = GraphConnector.get()
+    event = BackoffDetails(args=(graph,))
     GraphConnector._on_commit_backoff(event)
     assert init_driver.call_args_list == [call()]
     assert check_connectivity_and_authentication.call_args_list == [call()]
@@ -216,9 +216,9 @@ def test_on_commit_giveup(
         r"MATCH {{jinja_var}} RETURN $cypher_var;"
     )
 
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
     query = Query("match_something", template, {"jinja_var": "jjj"})
-    event = BackoffDetails(args=(connector, query), kwargs={"cypher_var": "ccc"})
+    event = BackoffDetails(args=(graph, query), kwargs={"cypher_var": "ccc"})
     GraphConnector._on_commit_giveup(event)
 
     assert caplog.messages[-1] == expected
@@ -226,9 +226,9 @@ def test_on_commit_giveup(
 
 def test_mocked_graph_commit_raises_error(mocked_graph: MockedGraph) -> None:
     mocked_graph.run.side_effect = Exception("query failed")
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
     with pytest.raises(Exception, match="query failed"):
-        connector.commit("RETURN 1;")
+        graph.commit("RETURN 1;")
 
 
 @pytest.mark.usefixtures("mocked_query_class")
@@ -253,6 +253,7 @@ def test_mocked_graph_fetch_extracted_items(mocked_graph: MockedGraph) -> None:
         query_string="my-query",
         stable_target_id=Identifier.generate(99),
         entity_type=["ExtractedFoo", "ExtractedBar", "ExtractedBatz"],
+        had_primary_source=None,
         skip=10,
         limit=100,
     )
@@ -260,7 +261,10 @@ def test_mocked_graph_fetch_extracted_items(mocked_graph: MockedGraph) -> None:
     assert mocked_graph.call_args_list[-1].args == (
         """\
 fetch_extracted_or_rule_items(
-    filter_by_query_string=True, filter_by_stable_target_id=True
+    filter_by_query_string=True,
+    filter_by_stable_target_id=True,
+    filter_by_reference_to_merged_item=False,
+    reference_field_name="hadPrimarySource",
 )""",
         {
             "labels": [
@@ -270,6 +274,7 @@ fetch_extracted_or_rule_items(
             ],
             "limit": 100,
             "query_string": "my-query",
+            "referenced_identifiers": None,
             "skip": 10,
             "stable_target_id": "bFQoRhcVH5DHV1",
         },
@@ -492,12 +497,13 @@ def test_fetch_extracted_items(
     limit: int,
     expected: dict[str, Any],
 ) -> None:
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
 
-    result = connector.fetch_extracted_items(
+    result = graph.fetch_extracted_items(
         query_string=query_string,
         stable_target_id=stable_target_id,
         entity_type=entity_type,
+        had_primary_source=None,
         skip=0,
         limit=limit,
     )
@@ -524,17 +530,21 @@ def test_mocked_graph_fetch_rule_items(mocked_graph: MockedGraph) -> None:
     ]
     graph = GraphConnector.get()
     result = graph.fetch_rule_items(
-        query_string="my-query",
-        stable_target_id=Identifier.generate(99),
-        entity_type=["AdditiveFoo", "SubtractiveBar", "PreventiveBatz"],
-        skip=10,
-        limit=100,
+        "my-query",
+        Identifier.generate(99),
+        ["AdditiveFoo", "SubtractiveBar", "PreventiveBatz"],
+        None,
+        10,
+        100,
     )
 
     assert mocked_graph.call_args_list[-1].args == (
         """\
 fetch_extracted_or_rule_items(
-    filter_by_query_string=True, filter_by_stable_target_id=True
+    filter_by_query_string=True,
+    filter_by_stable_target_id=True,
+    filter_by_reference_to_merged_item=False,
+    reference_field_name="hadPrimarySource",
 )""",
         {
             "labels": [
@@ -544,6 +554,7 @@ fetch_extracted_or_rule_items(
             ],
             "limit": 100,
             "query_string": "my-query",
+            "referenced_identifiers": None,
             "skip": 10,
             "stable_target_id": "bFQoRhcVH5DHV1",
         },
@@ -619,24 +630,18 @@ def test_fetch_rule_items(
     stable_target_id: str | None,
     expected: dict[str, Any],
 ) -> None:
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
 
-    result = connector.fetch_rule_items(
-        query_string=query_string,
-        stable_target_id=stable_target_id,
-        entity_type=None,
-        skip=0,
-        limit=1,
-    )
+    result = graph.fetch_rule_items(query_string, stable_target_id, None, None, 0, 1)
 
     assert result.one() == expected
 
 
 @pytest.mark.integration
 def test_fetch_rule_items_empty() -> None:
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
 
-    result = connector.fetch_rule_items(None, "thisIdDoesNotExist", None, 0, 1)
+    result = graph.fetch_rule_items(None, "thisIdDoesNotExist", None, None, 0, 1)
 
     assert result.one() == {"items": [], "total": 0}
 
@@ -1176,9 +1181,9 @@ def test_fetch_merged_items(  # noqa: PLR0913
     limit: int,
     expected: dict[str, Any],
 ) -> None:
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
 
-    result = connector.fetch_merged_items(
+    result = graph.fetch_merged_items(
         query_string=query_string,
         identifier=identifier,
         entity_type=entity_type,
@@ -1304,9 +1309,9 @@ exists_merged_item(node_labels=["MergedFoo", "MergedBar", "MergedBatz"])""",
 def test_graph_exists_merged_item(
     stable_target_id: Identifier, stem_types: list[str] | None, exists: bool
 ) -> None:
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
 
-    assert connector.exists_merged_item(stable_target_id, stem_types) == exists
+    assert graph.exists_merged_item(stable_target_id, stem_types) == exists
 
 
 @pytest.mark.usefixtures("mocked_query_class")
@@ -1653,10 +1658,11 @@ def test_mocked_graph_ingests_extracted_models(
         ],
     ]
 
+    dummy_items = list(dummy_data.values())
     graph = GraphConnector.get()
-    models = graph.ingest(list(dummy_data.values()))
+    graph.ingest(dummy_items)
 
-    assert models == list(dummy_data.values())
+    assert len(mocked_graph.call_args_list) == 18
 
 
 @pytest.mark.integration
@@ -1664,12 +1670,12 @@ def test_connector_flush_fails(monkeypatch: MonkeyPatch) -> None:
     settings = BackendSettings.get()
 
     monkeypatch.setattr(settings, "debug", False)
-    connector = GraphConnector.get()
+    graph = GraphConnector.get()
 
     with pytest.raises(
         MExError, match="database flush was attempted outside of debug mode"
     ):
-        connector.flush()
+        graph.flush()
 
 
 @pytest.mark.integration
@@ -1679,7 +1685,7 @@ def test_connector_flush(monkeypatch: MonkeyPatch) -> None:
 
     settings = BackendSettings.get()
     monkeypatch.setattr(settings, "debug", True)
-    connector = GraphConnector.get()
-    connector.flush()
+    graph = GraphConnector.get()
+    graph.flush()
 
     assert len(get_graph()) == 0